cvfspy/requirements.txt
# This template is a low-dependency template. 
# By default there is no requirements added here.
# Add the requirements you need to this file.
# or run `make init` to create this file automatically based on the template.
# You can also run `make switch-to-poetry` to use the poetry package manager.
genanki
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
newspaper3k
pydantic
wordfreq
flask
flask_cors